requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
email-validator==2.3.0

# AI and ML
anthropic>=0.42.0
openai==1.3.7
chromadb==0.4.18

# Document Processing
PyMuPDF>=1.24.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
pandas==2.1.3
openpyxl==3.1.2
python-docx==1.1.0
Pillow==10.1.0
matplotlib>=3.8.0

# Database and Storage
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0

# Redis for Caching and Rate Limiting (Tier 3)
redis>=5.0.0

# Text Processing
tiktoken==0.5.2
langchain-text-splitters==0.0.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
psutil==5.9.7
aiofiles==23.2.1

# Google Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
gspread==5.12.0

# PowerPoint Processing
python-pptx==0.6.23

# Export Tools
markdown==3.5.1
weasyprint==60.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
